backend/requirements.txt
# FastAPI core
fastapi==0.110.0
uvicorn==0.29.0

# HTTP async client
httpx==0.27.0

# Database / ORM (SQLite)
# SQLite is built-in — no external dependency needed

# Security / Authentication
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Environment variables
python-dotenv==1.0.1

# OpenAI SDK (Responses API)
openai==1.16.1

# Pydantic
pydantic==2.6.4
pydantic-settings==2.1.0

# CORS
starlette==0.36.3

# Data processing
numpy==1.26.4
scikit-learn==1.5.0
scipy==1.12.0
